refactor(cli): modularize CLI and default scaffolds to `vision-agents app`
Splits `cli/__main__.py` into a `main.py` group plus `init/` and `app/`
subpackages, each with its own `command.py`. Extracts scaffolding helpers
into `init/scaffold.py` and the project-config loader into `app/config.py`.
Updates the scaffolded README and the init "next steps" output to point
users at `vision-agents app run|serve` as the canonical entry.
