+++ agents-core/vision_agents/cli/init/__init__.py
@@ -1,0 +1,5 @@
+"""``vision-agents init`` subcommand."""
+
+from vision_agents.cli.init.command import init_cmd
+
+__all__ = ["init_cmd"]
