+++ agents-core/vision_agents/cli/init/templates/README.md.j2
@@ -14,12 +14,6 @@ Generated with `uvx vision-agents init`.
 3. Run the agent:
 
    ```bash
-   uv run python agent.py
-   ```
-
-   Or, from inside the project directory, use the `vision-agents` alias:
-
-   ```bash
    vision-agents app run     # single-call console
    vision-agents app serve   # HTTP server
    ```
