+++ agents-core/vision_agents/cli/init/templates/agent.py.j2
