+++ agents-core/vision_agents/cli/init/templates/env.example.j2
