+++ agents-core/vision_agents/cli/init/templates/gitignore.j2
