+++ agents-core/vision_agents/cli/init/templates/pyproject.toml.j2
