+++ agents-core/vision_agents/cli/init/templates/vision-agents.toml.j2
